moved pass or fail payload generator
