+++ src/services/CertificatePayloadGenerator.ts
@@ -4,7 +4,7 @@ import { ITestResult } from '../models/ITestResult';
 import { CertificatePayloadGeneratorIva } from './certificate-payload/CertificatePayloadGeneratorIva';
 import { CertificatePayloadGeneratorAdr } from './certificate-payload/CertificatePayloadGeneratorAdr';
 import { CertificatePayloadGeneratorRwt } from './certificate-payload/CertificatePayloadGeneratorRwt';
-import { CertificatePayloadGeneratorPassOrFail } from './CertificatePayloadGeneratorPassOrFail';
+import { CertificatePayloadGeneratorPassOrFail } from './certificate-payload/CertificatePayloadGeneratorPassOrFail';
 import { CertificatePayloadGeneratorMsva } from './certificate-payload/CertificatePayloadGeneratorMsva';
 
 // This is a facade
